pyproject.toml
[tool.poetry]
name = "fenixwbbot"
version = "0.1.0"
description = ""
authors = ["Vitaly Sokol <shanhozas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asyncio = "^3.4.3"
pymongo = "^4.3.3"
motor = "^3.1.1"
aiogram = {version = "^3.0.0b7", allow-prereleases = true}
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.0"
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
